requirements.txt
pyforms>=4.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.4.0
google-api-python-client>=2.0.0
PyQt5>=5.15.0
PyQtWebEngine>=5.15.0
